Create node_state_machine
A method of implementing a state machine design pattern using the game engine's node system. So far this state machine has been used to for our player character, chicken npc, and cow npc. It's really versatile, and I see myself using it for even more NPCs and characters in the future.